Several changes have been made...
1. Second txt file with users
2. Another window for managing multiple users (login, register)
3. Contact 'key' line has changed - userId added
4. Update function for txt file has changed - now, as an input 'add',
'modify', 'remove' is needed - thanks to this, one function allow to
update the file no matter what operation is used.
5. Vector containing contacts consists of only proper contacts (for
specific user). There was a need to change update function (previously
trunc) right now temporary file is used and later on renamed.
6. ContactID is taken from readContactsData (previously from vector) due
to the fact, the vector doesn't contain all contacts (only all available
for user)
7. Vector of all available contacts is updated after re-read of txt
file. Previously .push was used for all functions. The same is for a
vector of all users. Thanks to this, vector contains always the same
what txt file contains (if in example database would be edited from second source)